Fix NoSuchMethodError: JSONObject.put(String,Float) not available on Android 12
Android's org.json.JSONObject only provides put() overloads for:
- put(String, Double)
- put(String, Int)
- put(String, Long)
- put(String, Boolean)
- put(String, String)
- put(String, Object)

But NOT put(String, Float). This caused runtime crashes when
broadcasting audio_level and rtf metrics via WebSocket.

Fix: Cast Float values to Double before passing to JSONObject.put().

Locations fixed:
- broadcastAudioStatus(): audioLevel.toDouble()
- broadcastSpeechMessage(): rtf calculation uses toDouble()

This is a platform limitation, not a bug - Float must be boxed
to Double for JSON serialization on Android.

+++ app/src/main/java/com/uh/UhService.kt
@@ -798,6 +798,7 @@ class UhService : Service() {
         processingTimeMs: Long
     ) {
         try {
+            // Android 12 JSONObject.put() only accepts Double/Int, not Float
             val message = JSONObject().apply {
                 put("type", "speech")
                 put("text", text)
@@ -808,7 +809,7 @@ class UhService : Service() {
                 put("segment_end", endTime)
                 put("processing_time_ms", processingTimeMs)
                 put("audio_duration_ms", endTime - startTime)
-                put("rtf", processingTimeMs.toFloat() / (endTime - startTime).toFloat())
+                put("rtf", (processingTimeMs.toDouble() / (endTime - startTime).toDouble()))
             }
             
             webSocketServer?.broadcastMessage(message.toString())
@@ -828,10 +829,11 @@ class UhService : Service() {
             // Only broadcast if WebSocket server is running
             val server = webSocketServer ?: return
             
+            // Android 12 JSONObject.put() only accepts Double/Int, not Float
             val message = JSONObject().apply {
                 put("type", "audio_status")
                 put("listening", isListening)
-                put("audio_level", audioLevel)
+                put("audio_level", audioLevel.toDouble())
                 put("timestamp", System.currentTimeMillis())
             }
